swapped ports, moved spawn to seperate file
