+++ e2e-tests/studio.config.js
@@ -1,7 +1,7 @@
 /** @type {import('@yext/studio').StudioConfig} */
 export default {
   isPagesJSRepo: true,
-  port: 5173,
+  port: 5174,
   paths: {
     pages: "./src/templates",
   },
