+++ e2e-tests/tests/open-live-preview.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from "@playwright/test";
+import { expect, test } from "@playwright/test";
 import { studioTest } from "./infra/studioTest.js";
 
 studioTest(
