+++ packages/studio-plugin/src/createStudioPlugin.ts
@@ -1,6 +1,5 @@
 import { ConfigEnv, Plugin } from "vite";
 import { simpleGit } from "simple-git";
-import { spawn } from "node:child_process";
 import getStudioConfig from "./parsers/getStudioConfig";
 import ParsingOrchestrator, {
   createTsMorphProject,
@@ -18,6 +17,7 @@ import upath from "upath";
 import lodash from "lodash";
 import { UserConfig } from "vite";
 import { STUDIO_PROCESS_ARGS_OBJ } from "./constants";
+import { startPagesDevelopmentServer } from "./startPagesDevelopmentServer";
 
 /**
  * Handles server-client communication.
@@ -67,25 +67,7 @@ export default async function createStudioPlugin(
   );
 
   if (studioConfig.isPagesJSRepo) {
-    const pagesServer = spawn("npx", ["pages", "dev", "--open-browser=false"], {
-      stdio: ["ignore", "pipe", "pipe"],
-      env: {
-        ...process.env,
-      },
-      shell: true,
-    });
-
-    // stream pages server errors to parent process with a prefix
-    pagesServer.stderr.setEncoding("utf-8");
-    pagesServer.stderr.on("data", (data: string) => {
-      data.split("\n").forEach((line: string) => {
-        console.error("Pages Development Server |", line.trim());
-      });
-    });
-
-    pagesServer.addListener("exit", () => {
-      process.exit(-1);
-    });
+    startPagesDevelopmentServer()
   }
 
   return {
